Update main.gd to include an RST warning
Generally the main issue with having Intel RST enables is that ChimeraOS won't detect your drives when booting in UEFI mode. This adds an extra sentence to the end of the no drives detected error to inform users that such a feature must be set to off to detect said drives.

This will not solve every case of RST, however, because sometimes the installation media will detect itself as an installable drive. I am still unsure of how to warn them of this in that situation, but this should be enough to inform most users who have RST enabled.